Major codebase refactor
Reorganised files based on the MVVM pattern:
- models
- views
- viewmodels

Created:
- api_handler.py: handles calls to the steam api. SteamAPI class is stateful as it keeps track of api_key
- config_model.py: manages config file state, allows for reading/writing
- settings_model.py: manages settings file state, allows for reading/writing
- searched_users_model.py: manages searched users file state, allows for reading/writing
- core_viewmodel: acts as the application's intermediary between models and view
- Helpers folder for stateless files like file_manager.py or file_exporter.py

Deleted:
- core.py: replaced by core_viewmodel.py
- programsettings.py: deprecated due to introduction of settings, searched_users and config model files

To summarize, this feature completely refactors the architectural aspect of the project. Prior to these changes the project lacked structure, was very hard to track defects and the addition of new features became very difficult despite the fact that the application's size was relatively small.
Having adopted the MVVM pattern, the primary goal of setting foundations for a more modular and less coupled program has been achieved.